split server file and seeder file to avoide reseeding on connect
